remove unused exceljs from client bundle
