+++ models/server/createWorkbook.js
@@ -1,0 +1,5 @@
+import Excel from 'exceljs';
+
+export const createWorkbook = function() {
+  return new Excel.Workbook();
+};
